Fix numeric conversion safety and pre-commit latency (#336)
## Summary

- keep configured reconnection windows unsigned internally and saturate
deadlines at Chrono's upper bound instead of wrapping large values into
immediate expiry
- replace lossy percentile, duration, timestamp, capacity, metric,
schema, and test-fixture casts with exact checked or saturating
conversions
- render Prometheus integer metrics without an f64 round-trip and
enforce Clippy truncation/wrap/sign-loss lints across all root-package
targets
- restore the sub-second staged pre-commit path by directionally
classifying panic-policy diffs and overlapping worktree Git discovery
with policy setup

## Root cause and impact

`ServerConfig.reconnection_window` accepted `u64`, but
`ReconnectionManager` narrowed it with `as i64`. Values above `i64::MAX`
wrapped negative and minted already-expired reconnect tokens. The same
audit found other implicit numeric boundaries and an integer Prometheus
renderer that lost precision above 2^53.

The mandatory hook profile also showed 2.6–2.8 second Rust worktree
preflights. Added test guards unnecessarily loaded the full line-aware
panic scanner, while worktree status and Rust diff I/O ran serially.

## Safety and compatibility

Public `i64` reconnection APIs remain source-compatible. Wire schemas,
timeout values, dashboard percentile labels, dependency graphs, and the
P53/P56 hosted evidence selectors are unchanged. Added panic macros,
removed test guards, untracked production Rust, and Git errors remain
fail-closed.

## Validation

- `cargo fmt --check`
- `cargo clippy --locked --all-targets --all-features -- -D warnings`
- `cargo test --locked --all-features`
- `cargo deny --all-features check`
- all 313 `ci_config_tests` policy tests plus doc-consistency
policy/script suites
- CI-config, MSRV, docs, workflow hygiene, LLM-file, hook-readiness,
pre-commit, and pre-push gates
- five adversarial review passes; final settled review: zero findings
- final staged pre-commit: 500 ms during commit; five enforced profiles:
621–804 ms
- final worktree preflight: 809 ms; pre-push: 277–387 ms

Refs #213.
Refs #318.

<!-- CURSOR_SUMMARY -->
---

> [!NOTE]
> **Medium Risk**
> Touches reconnection token expiry and many metric/logging
conversions—user-visible reconnect behavior for extreme config
values—but behavior for normal windows is preserved and covered by
boundary tests; pre-commit changes are policy-only with directional
fixtures.
> 
> **Overview**
> **Numeric safety (issue #213):** Configured `u64` reconnection windows
no longer narrow to `i64` inside `ReconnectionManager`; expiry paths
stay unsigned and saturate at `DateTime::MAX_UTC` so extreme values do
not mint immediately expired tokens. Root `Cargo.toml` adds Clippy
`cast_possible_truncation`, `cast_possible_wrap`, and `cast_sign_loss`
(errors via `-D warnings`), with production and test code updated to use
`try_from`, saturating math, and integer percentile ranks instead of
lossy `as` casts. Prometheus counters/gauges render and scrape as exact
`u64` without an `f64` round-trip.
> 
> **Pre-commit latency (issue #318):** Worktree mode starts `git status`
and panic-policy `git diff` in parallel before policy setup; the Rust
panic gate parses zero-context hunks and only loads the full scanner
when a panic macro is added or a test-context guard is removed
(untracked Rust and Git failures still fail closed).
> 
> **CI:** `unused-deps.yml` always probes for `cargo-udeps 0.1.61` and
force-installs when missing or mismatched, avoiding a cache race where a
restored binary blocked reinstall.
> 
> <sup>Reviewed by [Cursor Bugbot](https://cursor.com/bugbot) for commit
0bb61a8a7042c389319828b882dfa0a3715e7cae. Bugbot is set up for automated
code reviews on this repo. Configure
[here](https://www.cursor.com/dashboard/bugbot).</sup>
<!-- /CURSOR_SUMMARY -->

+++ tests/v3_wire_properties.rs
@@ -290,7 +290,7 @@ proptest! {
             .and_then(Value::as_f64)
             .expect("float survives as a number");
 
-        let ulp_distance = (roundtripped.to_bits() as i64).abs_diff(float.to_bits() as i64);
+        let ulp_distance = roundtripped.to_bits().abs_diff(float.to_bits());
         prop_assert!(
             ulp_distance <= 4,
             "JSON float drift exceeded the pinned bound: {float} -> {roundtripped} ({ulp_distance} ULP)"
